Rename advancedlog package to log
