+++ analytics/legacy.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"time"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 )
 
 var (
