+++ cli/export.go
@@ -3,12 +3,11 @@ package cli
 import (
 	"encoding/json"
 
-	"gopkg.in/yaml.v2"
-
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/bitrise-io/bitrise/output"
 	"github.com/bitrise-io/go-utils/fileutil"
 	"github.com/urfave/cli"
+	"gopkg.in/yaml.v2"
 )
 
 func export(c *cli.Context) error {
