+++ cli/plugin.go
@@ -1,7 +1,7 @@
 package cli
 
 import (
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/urfave/cli"
 )
