+++ cli/plugin_install.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/bitrise-io/bitrise/plugins"
 	"github.com/urfave/cli"
 )
