+++ cli/run_util.go
@@ -12,11 +12,11 @@ import (
 	"strconv"
 	"time"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
-	"github.com/bitrise-io/bitrise/advancedlog/logwriter"
 	"github.com/bitrise-io/bitrise/analytics"
 	"github.com/bitrise-io/bitrise/bitrise"
 	"github.com/bitrise-io/bitrise/configs"
+	"github.com/bitrise-io/bitrise/log"
+	"github.com/bitrise-io/bitrise/log/logwriter"
 	"github.com/bitrise-io/bitrise/models"
 	"github.com/bitrise-io/bitrise/plugins"
 	"github.com/bitrise-io/bitrise/toolkits"
