+++ cli/setup.go
@@ -3,8 +3,8 @@ package cli
 import (
 	"os"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
 	"github.com/bitrise-io/bitrise/bitrise"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/urfave/cli"
 )
