+++ cli/version.go
@@ -2,10 +2,9 @@ package cli
 
 import (
 	"fmt"
-
 	"runtime"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/bitrise-io/bitrise/models"
 	"github.com/bitrise-io/bitrise/output"
 	"github.com/bitrise-io/bitrise/version"
