+++ log/corelog/console_logger.go
