+++ log/corelog/console_logger_test.go
