+++ log/corelog/json_logger_test.go
