+++ log/corelog/logger_test.go
@@ -4,7 +4,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/bitrise-io/bitrise/advancedlog/corelog"
+	"github.com/bitrise-io/bitrise/log/corelog"
 )
 
 func referenceTime() time.Time {
