+++ log/corelog/types.go
