+++ log/log.go
@@ -1,11 +1,11 @@
-package logger
+package log
 
 import (
 	"fmt"
 	"io"
 	"time"
 
-	"github.com/bitrise-io/bitrise/advancedlog/corelog"
+	"github.com/bitrise-io/bitrise/log/corelog"
 )
 
 type LoggerType corelog.LoggerType
