+++ log/log_functions.go
@@ -1,4 +1,4 @@
-package logger
+package log
 
 import (
 	"io"
