+++ log/logger.go
@@ -1,4 +1,4 @@
-package logger
+package log
 
 // Logger ...
 type Logger interface {
