+++ toolkits/command_runner.go
@@ -3,7 +3,7 @@ package toolkits
 import (
 	"fmt"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 	"github.com/bitrise-io/go-utils/command"
 	"github.com/bitrise-io/go-utils/errorutil"
 )
