+++ tools/filterwriter/filterwriter.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 	"time"
 
-	log "github.com/bitrise-io/bitrise/advancedlog"
+	"github.com/bitrise-io/bitrise/log"
 )
 
 // RedactStr ...
